Incluido confirmação do swal nas ações de exclusão e adição, e tratado mensagem de erro, quando o cpf consultado é invalido

+++ assets/js/requests.js
@@ -16,7 +16,11 @@ $(() => {
       })
       .fail((err) => {
         if (err.status === 500) {
-          swal('Oops', 'Ocorreu algum erro inesperado, tente novamente!', 'error');
+          return swal('Oops', 'Ocorreu algum erro inesperado, tente novamente!', 'error');
+        }
+
+        if (typeof err.responseJSON.msg !== 'undefined') {
+          return swal('Oops', err.responseJSON.msg, 'error');
         }
 
         alertError.addClass('d-none');
@@ -32,15 +36,27 @@ $(() => {
   const formAdd = $('#form-add');
 
   formAdd.submit(() => {
-    $.post(
-      formAdd.attr('action').slice(0, -1),
-      formAdd.serialize(),
-    )
-      .done(() => {
-        swal('Yeah', 'CPF adicionado a blacklist com sucesso!', 'success')
-          .then(() => window.location.reload());
-      })
-      .fail(() => swal('Oops', 'Este CPF já consta na blacklist!', 'error'));
+    swal({
+      title: 'Você tem certeza?',
+      text: 'Que deseja incluir este CPF a blacklist!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
+    })
+      .then((willDelete) => {
+        if (willDelete) {
+          $.post(
+            formAdd.attr('action').slice(0, -1),
+            formAdd.serialize(),
+          )
+            .done(() => {
+              swal('Yeah', 'CPF adicionado a blacklist com sucesso!', 'success')
+                .then(() => window.location.reload());
+            })
+            .fail(() => swal('Oops', 'Este CPF já consta na blacklist!', 'error'));
+        }
+      });
+
 
     return false;
   });
@@ -52,16 +68,26 @@ $(() => {
 
   formRemove.submit(() => {
     const inputCpf = $('#form-remove .input-cpf');
-
-    $.ajax({
-      url: formRemove.attr('action') + inputCpf.val(),
-      type: 'DELETE',
+    swal({
+      title: 'Você tem certeza?',
+      text: 'Que deseja remover este CPF da blacklist!',
+      icon: 'warning',
+      buttons: true,
+      dangerMode: true,
     })
-      .done(() => {
-        swal('Yeah', 'CPF removido da blacklist com sucesso!', 'success')
-          .then(() => window.location.reload());
-      })
-      .fail(() => swal('Oops', 'Este CPF já foi removido da blacklist!', 'error'));
+      .then((willDelete) => {
+        if (willDelete) {
+          $.ajax({
+            url: formRemove.attr('action') + inputCpf.val(),
+            type: 'DELETE',
+          })
+            .done(() => {
+              swal('Yeah', 'CPF removido da blacklist com sucesso!', 'success')
+                .then(() => window.location.reload());
+            })
+            .fail(() => swal('Oops', 'Este CPF já foi removido da blacklist!', 'error'));
+        }
+      });
 
     return false;
   });
